added Product to DataBase

+++ SignAndMarking/Controllers/FeatureController.cs
@@ -73,6 +73,10 @@ public class FeatureController : Controller
             return NotFound();
         }
         var feature = _db.Features.Find(Id);
+        if (feature == null)
+        {
+            return NotFound();
+        }
         return View(feature);
     }
 
@@ -86,6 +90,10 @@ public class FeatureController : Controller
             return NotFound();
         }
         var feature = _db.Features.Find(Id);
+        if (feature == null)
+        {
+            return NotFound();
+        }
         _db.Features.Remove(feature);
         _db.SaveChanges();
         return RedirectToAction("Index");
